requirements.txt
# Core dependencies
unstructured[pdf]>=0.16.0
llama-index>=0.12.0
llama-index-vector-stores-weaviate>=1.4.0
llama-index-embeddings-openai>=0.3.0
llama-index-llms-openai>=0.3.0
weaviate-client>=4.9.0
openai>=1.50.0
google-generativeai>=0.8.0
groq>=1.0.0
ragas>=0.2.0

# PDF processing
pdf2image>=1.17.0
pdfminer.six>=20231228
pikepdf>=9.0.0

# Frontend
streamlit>=1.40.0
streamlit-pdf-viewer>=0.0.15
fastapi>=0.115.0
uvicorn>=0.34.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.27.0
duckduckgo-search>=7.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0

# Evaluation
datasets>=3.0.0
langchain>=0.3.0
langchain-openai>=0.2.0
